+++ tests/__init__.py
@@ -1,0 +1,1 @@
+"""yaml_update_last_modified_date test files."""