+++ yaml_update_last_modified_date/__init__.py
@@ -1,0 +1,16 @@
+"""Set the last modified date in YAML front matter of edited markdown files."""
+
+# Keep these in sync with CHANGELOG.md and pyproject.toml.
+__name__ = "yaml-update-last-modified-date"
+__pkg_name__ = 'yaml_update_last_modified_date'
+__version__ = "1.1.1"  # FIXME: Update!
+
+
+# FIXME:
+# from importlib.metadata import version
+# __version__ = version(__package__)
+#
+
+# FIXME:
+# from loguru import logger
+# logger.disable(__pkg_name__)