+++ CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+<!-- FIXME: Add v1.2.1. Switch to python as basis for script. Will need to update homebrew and pre-commit hook. -->
+
+<!-- Keep latest version in sync with pyproject.toml and the main __init__.py file. -->
+
 ## [v1.1.1] - 2023-04-20
 
 ### Added