+++ CHANGELOG.md
@@ -19,9 +19,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-<!-- FIXME: Add v1.2.1. Switch to python as basis for script. Will need to update homebrew and pre-commit hook. -->
-
 <!-- Keep latest version in sync with pyproject.toml and the main __init__.py file. -->
+
+## [v2.0.0] - 2023-04-21  <!-- FIXME: Update release date. -->
+
+### Changed
+
+- Switch to python command-line utility instead of bash script for portability.
 
 ## [v1.1.1] - 2023-04-20
 
@@ -50,7 +54,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Quick Start section to README.md.
 -   This change log!
 
-[unreleased]: https://github.com/proinsias/yaml-update-last-modified-date/compare/v1.1.1...HEAD
+[unreleased]: https://github.com/proinsias/yaml-update-last-modified-date/compare/v2.0.0...HEAD
+[v2.0.0]: https://github.com/proinsias/yaml-update-last-modified-date/compare/v1.1.1...v2.0.0
 [v1.1.1]: https://github.com/proinsias/yaml-update-last-modified-date/compare/v1.1.0...v1.1.1
 [v1.1.0]: https://github.com/proinsias/yaml-update-last-modified-date/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/proinsias/yaml-update-last-modified-date/releases/tag/v1.0.0